pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "opencv-python>=4.12.0.88",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "uvicorn>=0.40.0",
]
